+++ _stbt/core.py
@@ -511,17 +511,20 @@ class MotionResult(object):
     * ``motion``: Boolean result, the same as evaluating `MotionResult` as a
       bool. That is, ``if result:`` will behave the same as
       ``if result.motion:``.
+    * ``region``: The region of the video frame that contained the motion.
+      `None` if no motion detected.
     """
-    def __init__(self, timestamp, motion):
+    def __init__(self, timestamp, motion, region=None):
         self.timestamp = timestamp
         self.motion = motion
+        self.region = region
 
     def __nonzero__(self):
         return self.motion
 
     def __repr__(self):
-        return "MotionResult(timestamp=%r, motion=%r)" % (
-            self.timestamp, self.motion)
+        return "MotionResult(timestamp=%r, motion=%r, region=%r)" % (
+            self.timestamp, self.motion, self.region)
 
 
 class OcrMode(IntEnum):
@@ -605,6 +608,47 @@ def new_device_under_test_from_config(
         restart_source, transformation_pipeline)
     return DeviceUnderTest(
         display=display, control=uri_to_remote(control_uri, display))
+
+
+def _pixel_bounding_box(img):
+    """
+    Find the smallest region that contains all the non-zero pixels in an image.
+
+    >>> _pixel_bounding_box(numpy.array([[0]], dtype=numpy.uint8))
+    >>> _pixel_bounding_box(numpy.array([[1]], dtype=numpy.uint8))
+    Region(x=0, y=0, right=1, bottom=1)
+    >>> _pixel_bounding_box(numpy.array([
+    ...     [0, 0, 0, 0],
+    ...     [0, 1, 1, 1],
+    ...     [0, 1, 1, 1],
+    ...     [0, 0, 0, 0],
+    ... ], dtype=numpy.uint8))
+    Region(x=1, y=1, right=4, bottom=3)
+    >>> _pixel_bounding_box(numpy.array([
+    ...     [0, 0, 0, 0, 0, 0],
+    ...     [0, 0, 0, 1, 0, 0],
+    ...     [0, 1, 0, 0, 0, 0],
+    ...     [0, 0, 0, 0, 1, 0],
+    ...     [0, 0, 1, 0, 0, 0],
+    ...     [0, 0, 0, 0, 0, 0]
+    ... ], dtype=numpy.uint8))
+    Region(x=1, y=1, right=5, bottom=5)
+    """
+    if len(img.shape) != 2:
+        raise ValueError("Single-channel image required.  Provided image has "
+                         "shape %r" % (img.shape,))
+
+    out = [None, None, None, None]
+
+    for axis in (0, 1):
+        flat = numpy.any(img, axis=axis)
+        indices = numpy.where(flat)[0]
+        if len(indices) == 0:
+            return None
+        out[axis] = indices[0]
+        out[axis + 2] = indices[-1] + 1
+
+    return Region.from_extents(*out)
 
 
 class DeviceUnderTest(object):
@@ -712,7 +756,9 @@ class DeviceUnderTest(object):
                         (template.name or template.image), first_pass_matched)
                     imglog.append(matches=result)
                     if grabbed_from_live:
-                        self._display.draw(result, None)
+                        self._display.draw(
+                            result, label="match(%r)" %
+                            os.path.basename(template.friendly_name))
                     yield result
 
             finally:
@@ -731,7 +777,8 @@ class DeviceUnderTest(object):
             result = self.match(
                 template, frame=sample, match_parameters=match_parameters,
                 region=region)
-            self._display.draw(result, None)
+            self._display.draw(result, label="match(%r)" %
+                               os.path.basename(template.friendly_name))
             yield result
 
     def detect_motion(self, timeout_secs=10, noise_threshold=None, mask=None):
@@ -782,25 +829,15 @@ class DeviceUnderTest(object):
             imglog.imwrite("absdiff_threshold", thresholded)
             imglog.imwrite("absdiff_threshold_erode", eroded)
 
-            motion = (cv2.countNonZero(eroded) > 0)
+            out_region = _pixel_bounding_box(eroded)
+            if out_region:
+                # Undo cv2.erode above:
+                out_region = out_region.extend(x=-1, y=-1)
 
-            # Visualisation: Highlight in red the areas where we detected motion
-            if motion:
-                with numpy_from_sample(sample) as frame:
-                    cv2.add(
-                        frame,
-                        numpy.multiply(
-                            numpy.ones(frame.shape, dtype=numpy.uint8),
-                            numpy.array((0, 0, 255), dtype=numpy.uint8),  # bgr
-                            dtype=numpy.uint8),
-                        mask=cv2.dilate(
-                            thresholded,
-                            cv2.getStructuringElement(
-                                cv2.MORPH_ELLIPSE, (3, 3)),
-                            iterations=1),
-                        dst=frame)
+            motion = bool(out_region)
 
-            result = MotionResult(sample.get_buffer().pts, motion)
+            result = MotionResult(sample.get_buffer().pts, motion, out_region)
+            self._display.draw(result, label="detect_motion()")
             debug("%s found: %s" % (
                 "Motion" if motion else "No motion", str(result)))
             yield result
@@ -1376,6 +1413,28 @@ def _mainloop():
               "is still alive!" if thread.isAlive() else "ok"))
 
 
+class _Annotation(namedtuple("_Annotation", "pts region label colour")):
+    MATCHED = (32, 0, 255)  # Red
+    NO_MATCH = (32, 255, 255)  # Yellow
+
+    @staticmethod
+    def from_result(result, label=""):
+        colour = _Annotation.MATCHED if result else _Annotation.NO_MATCH
+        return _Annotation(result.timestamp, result.region, label, colour)
+
+    def draw(self, img):
+        if not self.region:
+            return
+        cv2.rectangle(
+            img, (self.region.x, self.region.y),
+            (self.region.right, self.region.bottom), self.colour,
+            thickness=3)
+
+        # Slightly above the match annotation
+        label_loc = (self.region.x, self.region.y - 10)
+        _draw_text(img, self.label, label_loc, (255, 255, 255), font_scale=0.5)
+
+
 class Display(object):
     def __init__(self, user_source_pipeline, user_sink_pipeline,
                  mainloop, save_video="",
@@ -1393,7 +1452,7 @@ class Display(object):
 
         self.annotations_lock = threading.Lock()
         self.text_annotations = []
-        self.match_annotations = []
+        self.annotations = []
 
         self.restart_source_enabled = restart_source
 
@@ -1536,7 +1595,6 @@ class Display(object):
                             timeout_secs * Gst.SECOND))
                         return
 
-                sample = gst_sample_make_writable(sample)
                 try:
                     yield sample
                 finally:
@@ -1574,7 +1632,7 @@ class Display(object):
 
         self.last_sample.put_nowait(sample_or_exception)
 
-    def draw(self, obj, duration_secs):
+    def draw(self, obj, duration_secs=None, label=""):
         with self.annotations_lock:
             if isinstance(obj, str) or isinstance(obj, unicode):
                 obj = (
@@ -1582,9 +1640,10 @@ class Display(object):
                     ' ' + obj)
                 self.text_annotations.append(
                     {"text": obj, "duration": duration_secs * Gst.SECOND})
-            elif isinstance(obj, MatchResult):
-                if obj.timestamp is not None:
-                    self.match_annotations.append(obj)
+            elif hasattr(obj, "region") and hasattr(obj, "timestamp"):
+                annotation = _Annotation.from_result(obj, label=label)
+                if annotation.pts:
+                    self.annotations.append(annotation)
             else:
                 raise TypeError(
                     "Can't draw object of type '%s'" % type(obj).__name__)
@@ -1593,21 +1652,22 @@ class Display(object):
         # Calculate whether we need to draw any annotations on the output video.
         now = sample.get_buffer().pts
         texts = []
-        matches = []
+        annotations = []
         with self.annotations_lock:
             for x in self.text_annotations:
                 x.setdefault('start_time', now)
                 if now < x['start_time'] + x['duration']:
                     texts.append(x)
             self.text_annotations = texts[:]
-            for match_result in list(self.match_annotations):
-                if match_result.timestamp == now:
-                    matches.append(match_result)
-                if now >= match_result.timestamp:
-                    self.match_annotations.remove(match_result)
+            for annotation in list(self.annotations):
+                if annotation.pts == now:
+                    annotations.append(annotation)
+                if now >= annotation.pts:
+                    self.annotations.remove(annotation)
 
         sample = gst_sample_make_writable(sample)
         with numpy_from_sample(sample) as img:
+            # Text:
             _draw_text(
                 img, datetime.datetime.now().strftime("%H:%M:%S.%f")[:-4],
                 (10, 30), (255, 255, 255))
@@ -1616,16 +1676,10 @@ class Display(object):
                 age = float(now - x['start_time']) / (3 * Gst.SECOND)
                 color = (int(255 * max([1 - age, 0.5])),) * 3
                 _draw_text(img, x['text'], origin, color)
-            for match_result in matches:
-                if isinstance(match_result.image, numpy.ndarray):
-                    label = '<Custom Image>'
-                else:
-                    label = os.path.basename(match_result.image)
-                # Slightly above the match annotation
-                label_loc = (match_result.region.x, match_result.region.y - 10)
-                _draw_text(
-                    img, label, label_loc, (255, 255, 255), font_scale=0.5)
-                _draw_match(img, match_result.region, match_result.match, img)
+
+            # Regions:
+            for annotation in annotations:
+                annotation.draw(img)
 
         self.appsrc.props.caps = sample.get_caps()
         self.appsrc.emit("push-buffer", sample.get_buffer())
@@ -1733,6 +1787,9 @@ class Display(object):
 
 
 def _draw_text(numpy_image, text, origin, color, font_scale=1.0):
+    if not text:
+        return
+
     (width, height), _ = cv2.getTextSize(
         text, fontFace=cv2.FONT_HERSHEY_DUPLEX, fontScale=font_scale,
         thickness=1)
@@ -1743,16 +1800,6 @@ def _draw_text(numpy_image, text, origin, color, font_scale=1.0):
     cv2.putText(
         numpy_image, text, origin, cv2.FONT_HERSHEY_DUPLEX,
         fontScale=font_scale, color=color, lineType=cv2.CV_AA)
-
-
-def _draw_match(numpy_image, region, match_, output_image=None, thickness=3):
-    if output_image is None:
-        output_image = numpy_image.copy()
-    cv2.rectangle(
-        output_image, (region.x, region.y), (region.right, region.bottom),
-        (32, 0 if match_ else 255, 255),  # bgr
-        thickness=thickness)
-    return output_image
 
 
 class GObjectTimeout(object):
@@ -2079,16 +2126,18 @@ def _log_match_image_debug(imglog):
 
     for matched, position, _, level in imglog.data["pyramid_levels"]:
         template = imglog.images["level%d-template" % level]
-        imglog.imwrite("level%d-source_with_match" % level, _draw_match(
-            imglog.images["level%d-source" % level],
-            Region(x=position.x, y=position.y,
-                   width=template.shape[1], height=template.shape[0]),
-            matched, thickness=1))
+        imglog.imwrite("level%d-source_with_match" % level,
+                       imglog.images["level%d-source" % level],
+                       Region(x=position.x, y=position.y,
+                              width=template.shape[1],
+                              height=template.shape[0]),
+                       _Annotation.MATCHED if matched else _Annotation.NO_MATCH)
 
     for i, result in enumerate(imglog.data["matches"]):
-        imglog.imwrite("match%d-source_with_match" % i, _draw_match(
-            imglog.images["source"], result.region, result._first_pass_matched,  # pylint:disable=protected-access
-            thickness=1))
+        imglog.imwrite(
+            "match%d-source_with_match" % i, imglog.images["source"],
+            result.region, _Annotation.MATCHED if result._first_pass_matched  # pylint:disable=protected-access
+            else _Annotation.NO_MATCH)
 
     try:
         import jinja2
@@ -2581,6 +2630,9 @@ def _fake_frames_at_half_motion():
                 yield gst_sample_make_writable(
                     Gst.Sample.new(buf, Gst.Caps.from_string(
                         'video/x-raw,format=BGR,width=2,height=2'), None, None))
+
+        def draw(self, *_args, **_kwargs):
+            pass
 
     dut = DeviceUnderTest(display=FakeDisplay())
     dut.get_frame = lambda: None