+++ tests/test-motion.sh
@@ -77,6 +77,9 @@ test_detect_motion_reports_motion() {
 	for motion_result in detect_motion():
 	    assert bool(motion_result) == motion_result.motion
 	    if motion_result:
+	        # videotestsrc has motion in bottom right corner:
+	        assert motion_result.region == stbt.Region(
+	            240, 180, right=320, bottom=240)
 	        import sys
 	        sys.exit(0)
 	    else:
@@ -205,4 +208,26 @@ test_detect_motion_example_press_and_wait_for_no_motion() {
 	raise Exception("Timeout occured without any result reported.")
 	EOF
     stbt run -v test.py
+}
+
+test_detect_motion_visualisation() {
+    cat > detect_motion.py <<-EOF &&
+	for result in detect_motion():
+	    pass
+	EOF
+    mkfifo fifo || fail "Initial test setup failed"
+
+    stbt run -v \
+        --source-pipeline "multifilesrc location=$testdir/box-%05d.png loop=true" \
+        --sink-pipeline 'gdppay ! filesink location=fifo' \
+        detect_motion.py &
+    source_pid=$!
+    trap "kill $source_pid; rm fifo" EXIT
+
+    cat > verify.py <<-EOF &&
+	wait_for_match("$testdir/motion-visualisation.png")
+	EOF
+    stbt run -v --control none \
+        --source-pipeline 'filesrc location=fifo ! gdpdepay' \
+        verify.py
 }